requirements.txt
blinker==1.9.0
colorama==0.4.6
cycler==0.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fonttools==4.55.3
itsdangerous==2.2.0
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
pytz==2024.2
scipy==1.15.1
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.2
Werkzeug==3.1.3

# Agregar las siguientes dependencias necesarias
flask
pandas
matplotlib

